Reduce size of new image
The new, larger, image will test things more carefully, but it’s a little *too* big. This shrinks it to 40% in each dimension, which brings it down to just over 400 1K packets. This is good because it gets us above 256 packets, which means we test two byte packet numbers more fully.